% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_custom_ev.R
\name{add_custom_ev}
\alias{add_custom_ev}
\title{Add your own vehicle to the database}
\usage{
add_custom_ev(model, batterysize, consumption, range_km)
}
\arguments{
\item{model}{Modelname of the vehicle.}

\item{batterysize}{Batterysize in kWh (for EVs normally between 20 and 140)}

\item{consumption}{Consumption in kWh/100km (for EVs normally between 14 and 25)}

\item{range_km}{Official WLTP Range in km (for modern EVs normally between 250 and 900)}
}
\description{
Add your own vehicle to the database
}
