% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_buffer_topo.R
\name{calc_ev_reach}
\alias{calc_ev_reach}
\title{Topographic Buffer around the vehicle location (in coordinates)}
\usage{
calc_ev_reach(start_point, battery_kwh, consumption_base, dem)
}
\arguments{
\item{dem}{SpatRaster (von elevatr geladen)}

\item{lon}{Numeric longitude.}

\item{lat}{Numeric latitude.}

\item{model_name}{From the Database with corresponding range-value.}

\item{batterylevel}{The Batterylevel (in \%) of the vehicle (changes the total range), default = 100 \%.}
}
\description{
Topographic Buffer around the vehicle location (in coordinates)
}
