% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_buffer_topo.R
\name{calc_ev_topo_buffer}
\alias{calc_ev_topo_buffer}
\title{Topographic Buffer around the vehicle location (in coordinates)}
\usage{
calc_ev_topo_buffer(
  lon,
  lat,
  model_name,
  batterylevel = 100,
  temp_celsius = 20,
  z = 7,
  min_power = 50
)
}
\arguments{
\item{lon}{Numeric longitude.}

\item{lat}{Numeric latitude.}

\item{model_name}{From the Database with corresponding range-value.}

\item{batterylevel}{The Batterylevel (in \%) of the vehicle (changes the total range), default = 100 \%.}

\item{temp_celsius}{If the temperature is lower than 20°C it reduces the total range by 1.5 \% per degree (default = 20°C)}

\item{z}{Zoom-Level for DEM-resolution (6-8 recommended; default = 7)}

\item{min_power}{Minimum charging speed for the charging stations (default = 50 kW)}
}
\description{
Topographic Buffer around the vehicle location (in coordinates)
}
