% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csrf.R
\name{csrf}
\alias{csrf}
\title{Case specific random forests.}
\usage{
csrf(formula, training_data, test_data, params1 = list(), params2 = list())
}
\arguments{
\item{formula}{Object of class \code{formula} or \code{character} describing the model to fit.}

\item{training_data}{Training data of class \code{data.frame}.}

\item{test_data}{Test data of class \code{data.frame}.}

\item{params1}{Parameters for the proximity random forest grown in the first step.}

\item{params2Parameters}{for the prediction random forests grown in the second step.}
}
\value{
Predictions for the test dataset
}
\description{
Case specific random forests.
}
\examples{
TODO
}

