% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charging_stations.R
\name{fetch_chargers_ocm}
\alias{fetch_chargers_ocm}
\title{Fetch charging stations from OpenChargeMap}
\usage{
fetch_chargers_ocm(
  lat,
  lon,
  distance_km,
  min_power = 50,
  api_key = Sys.getenv("OCM_API_KEY")
)
}
\arguments{
\item{lat}{Numeric latitude.}

\item{lon}{Numeric longitude.}

\item{distance_km}{Distance in kilometers.}

\item{api_key}{OpenChargeMap API Key (4f60a79e-6773-4844-ba53-9671a50b3916)}
}
\value{
An sf object with charging station locations.
}
\description{
Fetch charging stations from OpenChargeMap
}
\keyword{internal}
