% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charging_stations.R
\name{fetch_charging_stations}
\alias{fetch_charging_stations}
\title{Internal helper-function to fetch charging stations and reduce the size of the "topo_buffer"-function}
\usage{
fetch_charging_stations(lat, lon, radius_km, api_key)
}
\description{
Internal helper-function to fetch charging stations and reduce the size of the "topo_buffer"-function
}
\keyword{internal}
