% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerminalNodeIDs.R
\name{getTerminalNodeIDs}
\alias{getTerminalNodeIDs}
\title{Get terminal node IDs of observations.}
\usage{
getTerminalNodeIDs(rf, dat)
}
\arguments{
\item{rf}{\code{ranger} object.}

\item{dat}{New dataset. Terminal node IDs for this dataset are obtained.}
}
\value{
Matrix with terminal nodeIDs for all observations in dataset and trees.
}
\description{
Get terminal node IDs of observations.
}
\examples{
library(ranger)
rf <- ranger(Species ~ ., data = iris, write.forest = TRUE)
getTerminalNodeIDs(rf, iris)
}

