% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.rangerts}
\alias{importance.rangerts}
\alias{importance}
\title{rangerts variable importance}
\usage{
\method{importance}{rangerts}(x, ...)
}
\arguments{
\item{x}{rangerts object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Variable importance measures.
}
\description{
Extract variable importance of rangerts object.
}
\seealso{
\code{\link{rangerts}}
}
\author{
Marvin N. Wright
}
