% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendRanges.R
\name{make_anchored_bins}
\alias{make_anchored_bins}
\title{Make bins extending away from a GRanges (which will be anchored at the start)}
\usage{
make_anchored_bins(anchor_gr, window_size, num_windows, ntop = NA)
}
\arguments{
\item{anchor_gr}{GRanges object to anchor on (this becomes the first bin)}

\item{window_size}{Size of the bins}

\item{num_windows}{Number of bins}

\item{ntop}{subset input GRanges to take top N regions (based on "score" column)}
}
\value{
New GRanges object with <anchor>---------> extended windows
}
\description{
Make bins extending away from a GRanges (which will be anchored at the start)
}
\examples{

gr100k <- make_anchored_bins(test_ranges, 10000, 10)

}
