% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetRanges.R
\name{mergeRanges}
\alias{mergeRanges}
\title{Merge Ranges from the given list of GRanges}
\usage{
mergeRanges(rlist, filterGenes, lengthCutoff = 2000)
}
\arguments{
\item{rlist}{A list of GRanges}

\item{lengthCutoff}{Only return regions upto this length (bp)}

\item{filterRanges}{GRanges, If provided , the regions overlapping with these regions are removed}
}
\value{
GRanges object
}
\description{
Merge Ranges from the given list of GRanges
}
\examples{

divergent <- subsetRange_byGeneDirection(testRanges, direction = "divergent")
divergent_merged <- mergeRanges(divergent)

}

