% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{parse.formula}
\alias{parse.formula}
\title{Parse formula}
\usage{
parse.formula(formula, data)
}
\arguments{
\item{formula}{Object of class \code{formula} or \code{character} describing the model to fit.}

\item{data}{Training data of class \code{data.frame}.}
}
\value{
Dataset including selected columns and interactions.
}
\description{
Parse formula and return dataset containing selected columns. 
Interactions are supported for numerical columns only. 
An interaction column is the product of all interacting columns.
}

