% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions.ranger}
\alias{predictions.ranger}
\title{Ranger predictions}
\usage{
\method{predictions}{ranger}(x, ...)
}
\arguments{
\item{x}{Ranger object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Predictions: Classes for Classification forests, Numerical values for Regressions forests and the estimated survival functions for all individuals for Survival forests.
}
\description{
Extract training data predictions of Ranger object.
}
\author{
Marvin N. Wright
}
\seealso{
\code{\link{ranger}}
}

