% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions.rangerts}
\alias{predictions.rangerts}
\title{rangerts predictions}
\usage{
\method{predictions}{rangerts}(x, ...)
}
\arguments{
\item{x}{rangerts object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Predictions: Classes for Classification forests, Numerical values for Regressions forests and the estimated survival functions for all individuals for Survival forests.
}
\description{
Extract training data predictions of rangerts object.
}
\seealso{
\code{\link{rangerts}}
}
\author{
Marvin N. Wright
}
