% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deforest.R
\name{print.deforest.ranger}
\alias{print.deforest.ranger}
\title{Print deforested ranger summary}
\usage{
\method{print}{deforest.ranger}(x, ...)
}
\arguments{
\item{x}{A \code{\link{deforest}} object (i.e., an object that inherits from
class \code{"deforest.ranger"}).}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print basic information about a deforested \code{\link{ranger}} object.
}
\note{
Many of the components of a typical \code{\link{ranger}} object are not 
available after deforestation and are instead replaced with \code{NA} (e.g., 
out-of-bag (OOB) predictions, variable importance scores (if requested), and 
OOB-based error metrics).
}
\seealso{
\code{\link{deforest}}.
}
\author{
Brandon M. Greenwell
}
