% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ranger.forest}
\alias{print.ranger.forest}
\title{Print Ranger forest}
\usage{
\method{print}{ranger.forest}(x, ...)
}
\arguments{
\item{x}{Object of class 'ranger.forest'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of Ranger forest object.
}
\author{
Marvin N. Wright
}

