% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.R
\name{print.ranger.prediction}
\alias{print.ranger.prediction}
\title{Print Ranger prediction}
\usage{
\method{print}{ranger.prediction}(x, ...)
}
\arguments{
\item{x}{Object of class 'ranger.prediction'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of Ranger prediction object.
}
\author{
Marvin N. Wright
}

