% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ranger}
\alias{print.ranger}
\title{Print Ranger}
\usage{
\method{print}{ranger}(x, ...)
}
\arguments{
\item{x}{Object of class 'ranger'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of Ranger object.
}
\author{
Marvin N. Wright
}
\seealso{
\code{\link{ranger}}
}

