% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rangerts.forest}
\alias{print.rangerts.forest}
\title{Print rangerts forest}
\usage{
\method{print}{rangerts.forest}(x, ...)
}
\arguments{
\item{x}{Object of class 'rangerts.forest'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of rangerts forest object.
}
\author{
Marvin N. Wright
}
