% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rangerts.prediction}
\alias{print.rangerts.prediction}
\title{Print rangerts prediction}
\usage{
\method{print}{rangerts.prediction}(x, ...)
}
\arguments{
\item{x}{Object of class 'rangerts.prediction'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of rangerts prediction object.
}
\author{
Marvin N. Wright
}
