% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rangerts}
\alias{print.rangerts}
\title{Print rangerts}
\usage{
\method{print}{rangerts}(x, ...)
}
\arguments{
\item{x}{Object of class 'rangerts'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of rangerts object.
}
\seealso{
\code{\link{rangerts}}
}
\author{
Marvin N. Wright
}
