% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{recode.factors}
\alias{recode.factors}
\title{Recode factor levels in data.frame}
\usage{
recode.factors(data, response)
}
\arguments{
\item{data}{Data frame with factor columns to recode}

\item{response}{Response variable}
}
\value{
Object of class \code{ranger.prediction} with elements
}
\description{
Prediction with new data and a saved forest from Ranger.
}
\author{
Marvin N. Wright
}
\seealso{
\code{\link{ranger}}
}

