% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetRanges.R
\name{subsetRange_byGeneDirection}
\alias{subsetRange_byGeneDirection}
\title{Filter a GRanges by the direction of genes}
\usage{
subsetRange_byGeneDirection(ranges, direction = "divergent")
}
\arguments{
\item{ranges}{A GRanges object to subset}

\item{direction}{Direction of Gene-pairs to subset. Options are 'convergent','divergent', or 'tandem'}
}
\value{
A list of two GRanges objects
}
\description{
Filter a GRanges by the direction of genes
}
\examples{

divergent <- subsetRange_byGeneDirection(testRanges, direction = "divergent")

}

