% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints.ranger}
\alias{timepoints}
\alias{timepoints.ranger}
\title{Ranger timepoints}
\usage{
\method{timepoints}{ranger}(x, ...)
}
\arguments{
\item{x}{Ranger Survival forest object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Unique death times
}
\description{
Extract unique death times of Ranger Survival forest
}
\author{
Marvin N. Wright
}
\seealso{
\code{\link{ranger}}
}

