% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints.rangerts.prediction}
\alias{timepoints.rangerts.prediction}
\title{rangerts timepoints}
\usage{
\method{timepoints}{rangerts.prediction}(x, ...)
}
\arguments{
\item{x}{rangerts Survival prediction object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Unique death times
}
\description{
Extract unique death times of rangerts Survival prediction object.
}
\seealso{
\code{\link{rangerts}}
}
\author{
Marvin N. Wright
}
