% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints.rangerts}
\alias{timepoints.rangerts}
\alias{timepoints}
\title{rangerts timepoints}
\usage{
\method{timepoints}{rangerts}(x, ...)
}
\arguments{
\item{x}{rangerts Survival forest object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Unique death times
}
\description{
Extract unique death times of rangerts Survival forest
}
\seealso{
\code{\link{rangerts}}
}
