% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeInfo.R
\name{treeInfo}
\alias{treeInfo}
\title{Tree information in human readable format}
\usage{
treeInfo(object, tree = 1)
}
\arguments{
\item{object}{\code{ranger} object.}

\item{tree}{Number of the tree of interest.}
}
\value{
A data.frame with the columns
\tabular{ll}{
      \code{A} \tab A. \cr
      \code{B} \tab B. \cr
      \code{C} \tab C. \cr
  }
}
\description{
Extract tree information of a \code{ranger} object.
}
\details{
Node and variable ID's are 0-indexed, i.e., node 0 is the root node. 
If the formula interface is used in the \code{ranger} call, the variable ID's are usually different to the original data used to grow the tree. 
Refer to the variable name instead to be sure.
}
\examples{
require(ranger)
rf <- ranger(Species ~ ., data = iris)
treeInfo(rf, 1)
}
\seealso{
\code{\link{ranger}}
}
\author{
Marvin N. Wright
}
