% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune.mtry}
\alias{tune.mtry}
\title{Tune mtry}
\usage{
tune.mtry(mtry, n = 1, ...)
}
\arguments{
\item{mtry}{Vector of \code{mtry} values to try.}

\item{n}{Number of replicates.}

\item{...}{Further arguments passed to Ranger.}
}
\value{
Prediction error for each tried \code{mtry} value.
}
\description{
Runs ranger for a vector of \code{mtry} values and returns the out of bag prediction error for each run.
For each \code{mtry} the analysis will be replicated \code{n} times.
}
\details{
The runs of ranger for different values of \code{mtry} are NOT based on the same bootstrap sample. 
Be sure to run enough replicates (\code{n}) before you trust the results.
}
\author{
Marvin N. Wright
}
\seealso{
\code{\link{ranger}}
}

