% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerminalNodeIDs.R
\name{getTerminalNodeIDs}
\alias{getTerminalNodeIDs}
\title{Get terminal node IDs (deprecated)}
\usage{
getTerminalNodeIDs(rf, dat)
}
\arguments{
\item{rf}{\code{ranger} object.}

\item{dat}{New dataset. Terminal node IDs for this dataset are obtained.}
}
\value{
Matrix with terminal nodeIDs for all observations in dataset and trees.
}
\description{
This function is deprecated. 
Please use predict() with \code{type = "terminalNodes"} instead.
This function calls predict() now.
}
\examples{
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
getTerminalNodeIDs(rf, iris)
}
