% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdoutRF.R
\name{holdoutRF}
\alias{holdoutRF}
\title{Hold-out random forests}
\usage{
holdoutRF(...)
}
\arguments{
\item{...}{All arguments are passed to \code{\link{ranger}()} (except \code{importance}, \code{case.weights}, \code{replace} and \code{holdout}.).}
}
\value{
Hold-out random forests with variable importance.
}
\description{
Grow two random forests on two cross-validation folds. 
Instead of out-of-bag data, the other fold is used to compute permutation importance.
Related to the novel permutation variable importance by Janitza et al. (2015).
}
\author{
Marvin N. Wright
}
\references{
Janitza, S., Celik, E. & Boulesteix, A.-L., (2015). A computationally fast variable importance test for random forest for high dimensional data, Technical Report 185, University of Munich, \url{https://epub.ub.uni-muenchen.de/25587}. \cr
}
\seealso{
\code{\link{ranger}}
}

