% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.ranger}
\alias{importance}
\alias{importance.ranger}
\title{ranger variable importance}
\usage{
\method{importance}{ranger}(x, ...)
}
\arguments{
\item{x}{ranger object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Variable importance measures.
}
\description{
Extract variable importance of ranger object.
}
\author{
Marvin N. Wright
}
\seealso{
\code{\link{ranger}}
}

