% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Range_Buffer_simple.R
\name{range_buffer_simple}
\alias{range_buffer_simple}
\title{Simple uniform Buffer around the vehicle location (in coordinates)}
\usage{
range_buffer_simple(lon, lat, model_name, batterylevel = 100)
}
\arguments{
\item{lon}{Numeric longitude.}

\item{lat}{Numeric latitude.}

\item{model_name}{From the Database with corresponding range-value.}

\item{batterylevel}{The Batterylevel (in \%) of the vehicle (changes the total range), default = 100 \%.}
}
\value{
Leaflet map.
}
\description{
Simple uniform Buffer around the vehicle location (in coordinates)
}
