% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buffer_simple.R
\name{range_buffer_simple}
\alias{range_buffer_simple}
\title{Simple uniform Buffer around the vehicle location (in coordinates)}
\usage{
range_buffer_simple(lon, lat, radius_meters = 1e+05)
}
\arguments{
\item{lon}{Numeric.}

\item{lat}{Numeric.}
}
\value{
Leaflet map.
}
\description{
Simple uniform Buffer around the vehicle location (in coordinates)
}
