% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints.ranger.prediction}
\alias{timepoints.ranger.prediction}
\title{Ranger timepoints}
\usage{
\method{timepoints}{ranger.prediction}(x, ...)
}
\arguments{
\item{x}{Ranger Survival prediction object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Unique death times
}
\description{
Extract unique death times of Ranger Survival prediction object.
}
\author{
Marvin N. Wright
}
\seealso{
\code{\link{ranger}}
}

